pyproject.toml
[tool.poetry]
name = "iceberg-sink-test"
version = "0.1.0"
description = ""
authors = ["risingwavelabs"]

[tool.poetry.dependencies]
python = "^3.10"
pyspark = { version = "3.4.1", extras = ["sql", "connect"] }
kafka-python = "2"
psycopg2 = "2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

